+++ packages/builder/src/schemas.zod.ts
@@ -1,5 +1,4 @@
 import { z } from 'zod';
-import importSpec from './steps/import';
 import { ethers } from 'ethers';
 
 /// ================================ INPUT CONFIG SCHEMAS ================================ \\\
@@ -17,7 +16,7 @@ const argsUnion = z.union([argtype, argtype2, argtype3, argtype4]);
 // <%=  string interpolation %>, step.names or property.names, packages:versions
 const interpolatedRegex = RegExp(/^<%=\s\w+.+[\w()[\]-]+\s%>$/, 'i');
 const stepRegex = RegExp(/^[\w-]+\.[.\w-]+$/, 'i');
-const packageRegex = RegExp(/^[\w-]+\.*:*[\w.-]+$/, 'i');
+const packageRegex = RegExp(/^[\w-]+\.*:*[\w.-@]+$/, 'i');
 const jsonAbiPathRegex = RegExp(/^(?!.*\.d?$).*\.json?$/, 'i');
 
 // This regex matches artifact names which are just capitalized words like solidity contract names
@@ -147,7 +146,7 @@ export const importSchema = z
     source: z.string().refine(
       (val) => !!val.match(packageRegex) || !!val.match(stepRegex) || !!val.match(interpolatedRegex),
       (val) => ({
-        message: `Source value: ${val} must match package format "package:version" or step format "import.Contract" or be an interpolated value`,
+        message: `Source value: ${val} must match package formats: "package:version" or "package:version@preset" or step format "import.Contract" or be an interpolated value`,
       })
     ),
   })
@@ -353,7 +352,7 @@ export const provisionSchema = z
     source: z.string().refine(
       (val) => !!val.match(packageRegex) || !!val.match(interpolatedRegex),
       (val) => ({
-        message: `Source value: ${val} must match package format "package:version" or be an interpolated value`,
+        message: `Source value: ${val} must match package formats: "package:version" or "package:version@preset" or be an interpolated value`,
       })
     ),
   })
@@ -398,6 +397,13 @@ export const provisionSchema = z
       })
       .deepPartial()
   );
+
+export const routerSchema = z.object({
+  contracts: z.array(z.string()),
+  from: z.string().optional(),
+  salt: z.string().optional(),
+  depends: z.array(z.string()).optional(),
+});
 
 /**
  * @internal NOTE: if you edit this schema, please also edit the constructor of ChainDefinition in 'definition.ts' to account for non-action components
@@ -463,14 +469,12 @@ export const chainDefinitionSchema = z
         /**
          * @internal
          */
-        import: z.custom<typeof importSpec>(),
+        import: z.record(importSchema),
+        provision: z.record(provisionSchema),
+        contract: z.record(contractSchema),
+        invoke: z.record(invokeSchema),
+        router: z.record(routerSchema),
+        // ... there may be others that come from plugins
       })
       .deepPartial()
   );
-
-export const routerSchema = z.object({
-  contracts: z.array(z.string()),
-  from: z.string().optional(),
-  salt: z.string().optional(),
-  depends: z.array(z.string()).optional(),
-});