+++ packages/website/next.config.js
@@ -2,13 +2,15 @@
 const nextConfig = {
   pageExtensions: ['ts', 'tsx', 'js', 'jsx', 'md', 'mdx'],
   reactStrictMode: true,
+  productionBrowserSourceMaps: true,
   webpack: (config) => {
     config.resolve.fallback = { fs: false, net: false, tls: false };
-    config.externals.push('pino-pretty', 'lokijs', 'encoding');
+    config.externals.push('pino-pretty', 'lokijs', 'encoding', 'solc');
     return config;
   },
   experimental: {
     mdxRs: true,
+    urlImports: ['https://unpkg.com'],
   },
 };