+++ packages/website/src/app/deploy/page.tsx
@@ -1,0 +1,14 @@
+import { Metadata } from 'next';
+import dynamic from 'next/dynamic';
+
+const NoSSR = dynamic(() => import('@/features/Deploy/SignTransactionsPage'), {
+  ssr: false,
+});
+
+export const metadata: Metadata = {
+  title: 'Cannon | Sign Transactions',
+};
+
+export default function SignTransactions() {
+  return <NoSSR />;
+}