+++ packages/website/src/types/StagedTransaction.ts
@@ -1,0 +1,6 @@
+import { SafeTransaction } from './SafeTransaction';
+
+export type StagedTransaction = {
+  txn: SafeTransaction;
+  sigs: string[];
+};