+++ packages/website/typings/ganache/ganache.d.ts
@@ -1,0 +1,1 @@
+declare module 'ganache/dist/web/ganache.min.js';