+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,38 @@
 # Changelog
 
+## [2.6.1](https://github.com/usecannon/cannon/compare/v2.6.0...v2.6.1) (2023-09-13)
+
+
+### Bug Fixes
+
+* gitHeads references from lerna ([ae76eae](https://github.com/usecannon/cannon/commit/ae76eaeda3e83ab14a09198449d0e0f096ee7839))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @usecannon/builder bumped from ^2.6.0 to ^2.6.1
+
+## [2.6.0](https://github.com/usecannon/cannon/compare/v2.5.4...v2.6.0) (2023-09-11)
+
+
+### Features
+
+* **router:** fix plugin loading ([f199216](https://github.com/usecannon/cannon/commit/f1992163722f28eb35b44e73c931e417b3c4ec1a))
+
+
+### Bug Fixes
+
+* **cli:** add error on upgrade-from when package not found ([#407](https://github.com/usecannon/cannon/issues/407)) ([2630d2c](https://github.com/usecannon/cannon/commit/2630d2ce9db57379aa0a473bac2e736160e72ed7))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @usecannon/builder bumped from ^2.5.4 to ^2.6.0
+
 ## [2.5.4](https://github.com/usecannon/cannon/compare/v2.5.3...v2.5.4) (2023-08-30)